fix: refactor controller folders
